Chatbot-Assistente-Cinema/requirements.txt
Flask
requests
python-dotenv
google-generativeai
Pillow